Cambios en categoria empresa y marca
Se puede agregar la imagen al momento de crear la entidad, marca terminada

+++ src/main/java/com/wisedevs/comv/services/EmpresaServiceImpl.java
@@ -1,15 +1,19 @@
 package com.wisedevs.comv.services;
 
+import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.validation.BindingResult;
+import org.springframework.web.multipart.MultipartFile;
 
+import com.wisedevs.comv.entities.Categoria;
 import com.wisedevs.comv.entities.Empresa;
+import com.wisedevs.comv.entities.Marca;
 import com.wisedevs.comv.repositories.IEmpresaRepository;
 
 @Service
@@ -21,33 +25,39 @@ public class EmpresaServiceImpl implements IEmpresaService{
 	@Autowired
 	private ICategoriaService CategoriaService;
 	
+	@Autowired
+	private ImagenServiceImpl imagenService;
+	
 	@Override
 	@Transactional
-	public ResponseEntity<String> save(Empresa empresa, String categoria) {
+	public ResponseEntity<String> save(Empresa empresa, String categoria,MultipartFile file)
+			throws IOException {
 	
 		
 		if (EmpresaRepository.existsByNombre(empresa.getNombre())) {
 			return new ResponseEntity<>("Ya existe una empresa con este nombre", HttpStatus.BAD_REQUEST);
 		}
 		
-		if (CategoriaService.existsByNombre(categoria)) {
-			empresa.setCategoria(CategoriaService.findbyNombre(categoria).get());
-			EmpresaRepository.save(empresa);
-			return new ResponseEntity<>("Empresa creada correctamente", HttpStatus.CREATED);
-		}else {
-			return new ResponseEntity<>("No existe una categoria con el nombre "+categoria, HttpStatus.BAD_REQUEST);
-
+		if(categoria!=null) {
+			Optional<Categoria> cat =  CategoriaService.findbyNombre(categoria);
+			if (cat.isPresent()) {
+				Categoria cate = cat.get();
+				empresa.setCategoria(cate);
+			}else {
+				return new ResponseEntity<>("No existe una categoria con el nombre "+categoria, HttpStatus.BAD_REQUEST);
+			}
 		}
+		EmpresaRepository.save(empresa);
+		imagenService.save(file, empresa.getNombre(), "Empresa");	
+		return new ResponseEntity<>("Empresa creada correctamente", HttpStatus.CREATED);
 		
 	}
 	
 	@Override
 	@Transactional
-	public ResponseEntity<String> actualizar(Empresa empresa, String categoria, BindingResult bindingResult) {
-		
-		if (bindingResult.hasErrors()) {
-            return new ResponseEntity<>("Campos con formato inválido", HttpStatus.BAD_REQUEST);
-        }
+	public ResponseEntity<String> actualizar(Empresa empresa, String categoria, 
+			MultipartFile file) throws IOException {
+
 		
 		if(empresa.getId()==null)
             return new ResponseEntity<>("Para actualizar info de una empresa se debe enviar su id", HttpStatus.BAD_REQUEST);
@@ -56,8 +66,12 @@ public class EmpresaServiceImpl implements IEmpresaService{
 		if (empresa.getNombre()!=null && EmpresaRepository.existsByNombre(empresa.getNombre())) {
 			return new ResponseEntity<>("Ya existe una empresa con este nombre", HttpStatus.BAD_REQUEST);
 		}
-		Empresa oldEmpresaInfo = EmpresaRepository.findById(empresa.getId()).get();
+		Optional<Empresa> OoldEmpresaInfo = EmpresaRepository.findById(empresa.getId());
 		
+		if(!OoldEmpresaInfo.isPresent()) {
+			return new ResponseEntity<>("No existe una empresa con este id", HttpStatus.BAD_REQUEST);
+		}
+		Empresa oldEmpresaInfo = OoldEmpresaInfo.get();
 		if(empresa.getNombre()==null) //si en el json no viene esta informacion, la info se toma de la oldempresa
 			empresa.setNombre(oldEmpresaInfo.getNombre());
 		if(!empresa.isMarca())
@@ -69,23 +83,38 @@ public class EmpresaServiceImpl implements IEmpresaService{
 		if(!empresa.isTurno())
 			empresa.setTurno(oldEmpresaInfo.isTurno());
 		
-		empresa.setBigImage(oldEmpresaInfo.getBigImage());
-		empresa.setMediumImage(oldEmpresaInfo.getMediumImage());//desde este controlador no se puede actualizar imagenes
-		empresa.setSmallImage(oldEmpresaInfo.getSmallImage());
 		
-		if (CategoriaService.existsByNombre(categoria)) {
-			empresa.setCategoria(CategoriaService.findbyNombre(categoria).get());
-			EmpresaRepository.save(empresa);
-			return new ResponseEntity<>("Empresa actualizada correctamente", HttpStatus.CREATED);
-		}else {
-			return new ResponseEntity<>("No existe una categoria con el nombre "+categoria, HttpStatus.BAD_REQUEST);
+		empresa.setBigImage(oldEmpresaInfo.getBigImage());
+		empresa.setMediumImage(oldEmpresaInfo.getMediumImage());
+		empresa.setSmallImage(oldEmpresaInfo.getSmallImage());	
 
+		if(categoria!=null) {
+			Optional<Categoria> cat = CategoriaService.findbyNombre(categoria);
+		
+			if (cat.isPresent()) {
+				empresa.setCategoria(cat.get());
+			}else {
+				return new ResponseEntity<>("No existe una categoria con el nombre "+categoria, HttpStatus.BAD_REQUEST);
+			}
 		}
+		EmpresaRepository.save(empresa);
+		if(file!=null && !file.isEmpty()) {
+			imagenService.delete(empresa.getId(), "Empresa");
+			imagenService.save(file, empresa.getNombre(), "Empresa");	
+		}
+		
+		
+		return new ResponseEntity<>("Empresa actualizada correctamente", HttpStatus.CREATED);
 	}
 
 	@Override
 	@Transactional
 	public void delete(Empresa empresa) {
+		imagenService.delete(empresa.getId(), "Empresa");
+		for(Marca m: empresa.getMarcas()) {
+			m.setEmpresas(null);
+		}
+		empresa.setMarcas(null);
 		EmpresaRepository.delete(empresa);
 		
 	}
@@ -93,7 +122,18 @@ public class EmpresaServiceImpl implements IEmpresaService{
 	@Override
 	@Transactional
 	public void deleteById(Long id) {
-		EmpresaRepository.deleteById(id);
+		
+		Optional<Empresa> empresaOp = EmpresaRepository.findById(id);
+		if(empresaOp.isPresent()) {
+			imagenService.delete(id, "Empresa");
+			Empresa empresa = empresaOp.get();
+			for(Marca m: empresa.getMarcas()) {
+				m.setEmpresas(null);
+			}
+			empresa.setMarcas(null);
+			EmpresaRepository.deleteById(id);
+		}
+		
 		
 	}
